requirements.txt
environs==14.3.0
python-telegram-bot==13.5
redis==6.4.0
